+++ CHANGELOG.md
@@ -5,11 +5,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [2.21.1] - 2019-10-02
 ### Added
 - ObjectType cache in connection.
 
 ## Changed
 - ObjectType.Close became unexported.
+- ODPI-C v3.2.2
+- de-embed conn from Queue.
 
 ## [2.20.0] - 2019-09-04
 ### Added